Remove debugger (#22502)
- Remove Debugger in the python extension, but only the fact to create a
configuration using 'python' type.
- It is still possible to debug using 'python' type config.
- Add extension "Python Debugger" to extension pack
