+++ gulpfile.js
@@ -98,7 +98,7 @@ async function addExtensionPackDependencies() {
     // extension dependencies need not be installed during development
     const packageJsonContents = await fsExtra.readFile('package.json', 'utf-8');
     const packageJson = JSON.parse(packageJsonContents);
-    packageJson.extensionPack = ['ms-python.vscode-pylance'].concat(
+    packageJson.extensionPack = ['ms-python.vscode-pylance', 'ms-python.debugpy'].concat(
         packageJson.extensionPack ? packageJson.extensionPack : [],
     );
     // Remove potential duplicates.
