--- src/client/debugger/extension/configuration/providers/pidAttach.ts
@@ -1,29 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-'use strict';
-
-import { DebugConfigStrings } from '../../../../common/utils/localize';
-import { MultiStepInput } from '../../../../common/utils/multiStepInput';
-import { sendTelemetryEvent } from '../../../../telemetry';
-import { EventName } from '../../../../telemetry/constants';
-import { DebuggerTypeName } from '../../../constants';
-import { AttachRequestArguments } from '../../../types';
-import { DebugConfigurationState, DebugConfigurationType } from '../../types';
-
-export async function buildPidAttachConfiguration(
-    _input: MultiStepInput<DebugConfigurationState>,
-    state: DebugConfigurationState,
-): Promise<void> {
-    const config: Partial<AttachRequestArguments> = {
-        name: DebugConfigStrings.attachPid.snippet.name,
-        type: DebuggerTypeName,
-        request: 'attach',
-        processId: '${command:pickProcess}',
-        justMyCode: true,
-    };
-    sendTelemetryEvent(EventName.DEBUGGER_CONFIGURATION_PROMPTS, undefined, {
-        configurationType: DebugConfigurationType.pidAttach,
-    });
-    Object.assign(state.config, config);
-}
