+++ src/client/debugger/extension/serviceRegistry.ts
@@ -12,11 +12,6 @@ import { DebugSessionLoggingFactory } from './adapter/logging';
 import { OutdatedDebuggerPromptFactory } from './adapter/outdatedDebuggerPrompt';
 import { AttachProcessProviderFactory } from './attachQuickPick/factory';
 import { IAttachProcessProviderFactory } from './attachQuickPick/types';
-import { PythonDebugConfigurationService } from './configuration/debugConfigurationService';
-import { DynamicPythonDebugConfigurationService } from './configuration/dynamicdebugConfigurationService';
-import { LaunchJsonCompletionProvider } from './configuration/launch.json/completionProvider';
-import { InterpreterPathCommand } from './configuration/launch.json/interpreterPathCommand';
-import { LaunchJsonUpdaterService } from './configuration/launch.json/updaterService';
 import { AttachConfigurationResolver } from './configuration/resolvers/attach';
 import { DebugEnvironmentVariablesHelper, IDebugEnvironmentVariablesService } from './configuration/resolvers/helper';
 import { LaunchConfigurationResolver } from './configuration/resolvers/launch';
@@ -25,35 +20,9 @@ import { DebugCommands } from './debugCommands';
 import { ChildProcessAttachEventHandler } from './hooks/childProcessAttachHandler';
 import { ChildProcessAttachService } from './hooks/childProcessAttachService';
 import { IChildProcessAttachService, IDebugSessionEventHandlers } from './hooks/types';
-import {
-    IDebugAdapterDescriptorFactory,
-    IDebugConfigurationService,
-    IDebugSessionLoggingFactory,
-    IDynamicDebugConfigurationService,
-    IOutdatedDebuggerPromptFactory,
-} from './types';
+import { IDebugAdapterDescriptorFactory, IDebugSessionLoggingFactory, IOutdatedDebuggerPromptFactory } from './types';
 
 export function registerTypes(serviceManager: IServiceManager): void {
-    serviceManager.addSingleton<IExtensionSingleActivationService>(
-        IExtensionSingleActivationService,
-        LaunchJsonCompletionProvider,
-    );
-    serviceManager.addSingleton<IExtensionSingleActivationService>(
-        IExtensionSingleActivationService,
-        InterpreterPathCommand,
-    );
-    serviceManager.addSingleton<IExtensionSingleActivationService>(
-        IExtensionSingleActivationService,
-        LaunchJsonUpdaterService,
-    );
-    serviceManager.addSingleton<IDebugConfigurationService>(
-        IDebugConfigurationService,
-        PythonDebugConfigurationService,
-    );
-    serviceManager.addSingleton<IDynamicDebugConfigurationService>(
-        IDynamicDebugConfigurationService,
-        DynamicPythonDebugConfigurationService,
-    );
     serviceManager.addSingleton<IChildProcessAttachService>(IChildProcessAttachService, ChildProcessAttachService);
     serviceManager.addSingleton<IDebugSessionEventHandlers>(IDebugSessionEventHandlers, ChildProcessAttachEventHandler);
     serviceManager.addSingleton<IDebugConfigurationResolver<LaunchRequestArguments>>(
