+++ src/client/telemetry/constants.ts
@@ -44,8 +44,6 @@ export enum EventName {
     DEBUGGER = 'DEBUGGER',
     DEBUGGER_ATTACH_TO_CHILD_PROCESS = 'DEBUGGER.ATTACH_TO_CHILD_PROCESS',
     DEBUGGER_ATTACH_TO_LOCAL_PROCESS = 'DEBUGGER.ATTACH_TO_LOCAL_PROCESS',
-    DEBUGGER_CONFIGURATION_PROMPTS = 'DEBUGGER.CONFIGURATION.PROMPTS',
-    DEBUGGER_CONFIGURATION_PROMPTS_IN_LAUNCH_JSON = 'DEBUGGER.CONFIGURATION.PROMPTS.IN.LAUNCH.JSON',
 
     // Python testing specific telemetry
     UNITTEST_CONFIGURING = 'UNITTEST.CONFIGURING',
